fix: prevents system adding the object back after force delete #775
This adresses an issue when deleting a post with the force-delete option with an attached featured image. Due to the process of deleting a post, algolia will add back the post when meta data is updated and in this process re-create the post.

This fix is keeping track of what has been deleted from the index and aborts all following actions taken to the post id. This will also affect number of re-index calls consumed while deleting a post.